Some more float/int cleanup

+++ src/worldmap/level_dot.cpp
@@ -130,7 +130,7 @@ LevelDot::draw_hover(DrawingContext& gc)
   if (accessible())
     {
       int length = Fonts::pingus_small.bounding_rect(0, 0, _(get_plf().get_levelname())).get_width() / 2;
-      int realpos = static_cast<int>(gc.world_to_screen(Vector2i(pos.x, pos.y)).x);
+      int realpos = gc.world_to_screen(Vector2i(static_cast<int>(pos.x), static_cast<int>(pos.y))).x;
       if (realpos - length < 0)
         pos_correction = realpos - length;
       else if (realpos + length > gc.get_width())
@@ -145,7 +145,7 @@ LevelDot::draw_hover(DrawingContext& gc)
   else
     {
       int length  = Fonts::pingus_small.bounding_rect(0, 0, _("locked")).get_width() / 2;
-      int realpos = static_cast<int>(gc.world_to_screen(Vector2i(pos.x, pos.y)).x);
+      int realpos = gc.world_to_screen(Vector2i(static_cast<int>(pos.x), static_cast<int>(pos.y))).x;
       if (realpos - length < 0)
         pos_correction = realpos - length;
       else if (realpos + length > gc.get_width())
